requirements.txt
Django==1.5.4
Fabric==1.8.0
Markdown==2.3.1
South==0.8.2
argparse==1.2.1
astroid==1.0.1
coverage==3.7
distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-nose==1.2
django-oauth2-provider==0.2.6
django-registration==1.0
django-toolbelt==0.0.1
djangorestframework==2.3.8
ecdsa==0.9
factory-boy==2.2.1
gunicorn==18.0
logilab-common==0.60.0
nose==1.3.0
paramiko==1.12.0
psycopg2==2.5.1
pycrypto==2.6.1
pylint==1.0.0
shortuuid==0.4
static==0.4
wsgiref==0.1.2
